chore: move statics from mocks to fixtures
