+++ test/__fixtures__/statics.ts
